style: 수익 출력값 예시와 동일하게 정수형으로 변경

+++ src/main/java/blackjack/view/OutputView.java
@@ -49,9 +49,9 @@ public class OutputView {
 
     public static void printAllEarnings(DealerResult dealerResult, List<ScoreResult> scoreResults) {
         System.out.println("## 최종 수익");
-        System.out.println(dealerResult.getName() + ": " + dealerResult.calculateEarnings());
+        System.out.println(dealerResult.getName() + ": " + (int) dealerResult.calculateEarnings());
         for (ScoreResult scoreResult : scoreResults) {
-            System.out.println(scoreResult.getPlayerName() + ": " + scoreResult.calculateEarnings());
+            System.out.println(scoreResult.getPlayerName() + ": " + (int) scoreResult.calculateEarnings());
         }
     }
 }
